Move intermediateToolingModelProvider impl to core
